changed to distanceCuts type to vector<double>

+++ LinearizedTrackFit/interface/BuildMatrix.h
@@ -16,7 +16,7 @@
 namespace LinearFit {
 
   void buildMatrix(const TString & inputFileName, const double & eventsFractionStart, const double & eventsFractionEnd,
-      const std::unordered_map<std::string, std::unordered_set<int> > & requiredLayers, const std::vector<float> distanceCuts,
+      const std::unordered_map<std::string, std::unordered_set<int> > & requiredLayers, const std::vector<double> distanceCuts,
       const std::vector<std::string> & inputVarNames, const std::vector<std::string> & inputTrackParameterNames, bool singleModules,
       const GeometricIndex::GeometricIndexConfiguration & gic)
   {
